Fix "asdf: command not found" error in configure script
The script was renaming the asdf binary as /home/user/.local/bin and putting it in $HOME/.local instead of putting it in the folder $HOME/.local/bin. This happens when the .local/bin doesn't exist. So to fix this, now the script checks if the directory exists, and if not, it's created